% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_comparison}
\alias{add_comparison}
\title{add_comparison}
\usage{
add_comparison(user = NA, pageid_1 = NA, pageid_2 = NA,
  name_1 = NA, name_2 = NA, more_left = NA, time = NA,
  party = NA, con = NULL)
}
\description{
add_comparison
}
